@@ -16,18 +16,17 @@
         }
     ],
     "require": {
-        "php": "^7.4|^8.0",
+        "php": "^8.0.0|^8.1.0",
         "spatie/laravel-package-tools": "^1.4.3",
-        "illuminate/contracts": "^8.37",
-        "bluefyn-international/sidekick": "^1.0",
+        "illuminate/contracts": "^9.0",
+        "bluefyn-international/sidekick": "^3.0",
         "brick/money": "^0.5.0",
         "laravelcollective/html": "^6.2"
     },
     "require-dev": {
         "barryvdh/laravel-ide-helper": "^2.8",
-        "brianium/paratest": "^6.2",
-        "nunomaduro/collision": "^5.3",
-        "orchestra/testbench": "^6.15",
+        "nunomaduro/collision": "^6.0",
+        "orchestra/testbench": "^7.0",
         "phpunit/phpunit": "^9.3",
         "spatie/laravel-ray": "^1.9",
         "vimeo/psalm": "^4.4"
